Moved assertSetupCompleted out into auth.setup module
